Repairing Alpha warnings, see the changelog for 2.11
git-svn-id: https://stealth.svn.sourceforge.net/svnroot/yodl/trunk@124 7de777c3-3b1c-0410-83c5-80958bd9f4b1

+++ yodl/src/message/messagesetverbosity.c
@@ -11,7 +11,7 @@
 
 void message_setverbosity(int mode, char *arg)
 {
-    size_t mask = MSG_NONE;
+    unsigned mask = MSG_NONE;
 
     if (sscanf(arg, "0x%x", (unsigned *)&mask) != 1)
     {
